Reorganize and add new withSession
