+++ src/store/memory.js
@@ -1,11 +1,7 @@
-/* eslint-disable class-methods-use-this */
-
-const Store = require('./store');
-
+const Store = require('../store');
 const MemoryStoreSession = {};
 
-class MemoryStore extends Store {
-  // eslint-disable-next-line no-useless-constructor
+export default class MemoryStore extends Store {
   constructor() {
     super();
     this.sessions = MemoryStoreSession;
@@ -67,5 +63,3 @@ class MemoryStore extends Store {
     return Promise.resolve();
   }
 }
-
-module.exports = MemoryStore;
