+++ src/withSession.jsx
@@ -1,0 +1,32 @@
+import { applySession, getOptions } from './core';
+
+function applyHOC(Page, hookName, options) {
+  function WithSession(props) {
+    return <Page {...props} />;
+  }
+  function getDisplayName(WrappedComponent) {
+    return WrappedComponent.displayName || WrappedComponent.name || 'Component';
+  }
+  WithSession.displayName = `WithSession(${getDisplayName(Page)})`;
+  WithSession[hookName] = async (pageCtx) => {
+    const ctx = 'Component' in pageCtx ? pageCtx.ctx : pageCtx;
+    if (typeof window === 'undefined') {
+      const { req, res } = ctx;
+      await applySession(req, res, options);
+    }
+    return Page[hookName](ctx);
+  };
+}
+
+export default function withSession(handler, opts) {
+  const options = getOptions(opts);
+  // Page Components
+  if (handler.getServerProps) return applyHOC(handler, 'getServerProps', options);
+  if (handler.unstable_getServerProps) return applyHOC(handler, 'getServerProps', options);
+  if (handler.getInitialProps) return applyHOC(handler, 'getInitialProps', options);
+  // API Routes
+  return async function WithSession(req, res) {
+    await applySession(req, res, options);
+    return handler(req, res);
+  };
+}
